Refractoring, Timestamp handling, Previous messages
I refractored the activities, so the a class called UiHandler handles
the navigation drawer and other graphical elements that is like on both
activities.
Also timestamps has been added and are shown different based upon how
long since the message was sent. Also the previous messages is
functional with some data